update helper time formatter
